requirements.txt
openai==2.1.0
streamlit==1.50.0
python-dotenv==1.1.1
IPython==9.6.0
google-generativeai==0.8.5
gradio==5.49.0
beautifulsoup4==4.14.2